Implemented new paths for name service in names.ts. Added endpoints to register names, set text records and resolve names. Improved the buildTypedData function to accept custom types and implemented the buildZyfiSetTextRecordRequest function. Additional validations are performed for input data and error handling logic is adjusted.
