+++ clk-gateway/src/index.ts
@@ -24,12 +24,6 @@ import {
   isParsableError,
   isOffchainLookupError,
   safeUtf8Decode,
-  getMessageHash,
-  validateSignature,
-  getDecodedToken,
-  checkUserByEmail,
-  asyncHandler,
-  buildTypedData,
 } from "./helpers";
 import admin from "firebase-admin";
 import {
@@ -44,8 +38,6 @@ import {
   nodleNSDomain,
   parentTLD,
   zyfiSponsoredUrl,
-  l1Provider,
-  nodleNameServiceContract,
   buildZyfiRegisterRequest,
 } from "./setup";
 import { getBatchInfo, fetchZyfiSponsored } from "./helpers";
@@ -506,239 +498,6 @@ app.post(
     }
   }
 );
-
-app.post(
-  "/name/register",
-  [
-    body("name")
-      .isLowercase()
-      .withMessage("Name must be a lowercase string")
-      .isFQDN()
-      .withMessage("Name must be a fully qualified domain name")
-      .custom((name) => {
-        const [sub, domain, tld] = name.split(".");
-        if (![`${clickNSDomain}.${parentTLD}`, `${nodleNSDomain}.${parentTLD}`].includes(`${domain}.${tld}`)) {
-          return false;
-        }
-
-        const subHash = keccak256(toUtf8Bytes(sub));
-        if (reservedHashes.includes(subHash)) {
-          return false;
-        }
-
-        return true;
-      })
-      .withMessage("Invalid domain or tld or reserved subdomain")
-      .custom((name) => {
-        const [sub] = name.split(".");
-        if (sub.length < 5) {
-          return false;
-        }
-        return true;
-      })
-      .withMessage(
-        "Current available subdomain names are limited to those with at least 5 characters"
-      ),
-    body("signature")
-      .isString()
-      .custom((signature) => {
-        return isHexString(signature);
-      })
-      .withMessage("Signature must be a hex string"),
-    body("owner")
-      .isString()
-      .custom((owner) => {
-        return isAddress(owner);
-      })
-      .withMessage("Owner must be a valid Ethereum address"),
-    body("email").isEmail().withMessage("Email must be a valid email address"),
-  ],
-  asyncHandler(async (req: Request, res: Response) => {
-    const decodedToken = await getDecodedToken(req);
-
-    const result = validationResult(req);
-    if (!result.isEmpty()) {
-      throw new HttpError(
-        result
-          .array()
-          .map((error) => error.msg)
-          .join(", "),
-        400
-      );
-    }
-    const data = matchedData(req);
-    const [name,sub] = data.name.split(".");
-    const owner = getAddress(data.owner);
-
-    const typedData = buildTypedData({
-      name: data.name,
-      email: data.email,
-    });
-
-    const isValidSignature = validateSignature({
-      typedData,
-      signature: data.signature,
-      expectedSigner: owner,
-    });
-
-    if (!isValidSignature) {
-      throw new HttpError("Invalid signature", 403);
-    }
-
-    let response;
-    if (zyfiSponsoredUrl) {
-      const zyfiRequest = buildZyfiRegisterRequest(owner, name, sub);
-      const zyfiResponse = await fetchZyfiSponsored(zyfiRequest);
-      console.log(`ZyFi response: ${JSON.stringify(zyfiResponse)}`);
-
-      await admin.auth().revokeRefreshTokens(decodedToken.uid);
-
-      response = await l2Wallet.sendTransaction(zyfiResponse.txData);
-    } else {
-      await admin.auth().revokeRefreshTokens(decodedToken.uid);
-
-      response = await clickNameServiceContract.register(owner, sub);
-    }
-
-    const receipt = await response.wait();
-    if (receipt.status !== 1) {
-      throw new Error("Transaction failed");
-    }
-
-    await admin
-      .auth()
-      .setCustomUserClaims(decodedToken.uid, { subDomain: sub });
-    res.status(200).send({
-      txHash: receipt.hash,
-    });
-  })
-);
-
-app.post(
-  "/name/register/message",
-  [
-    body("name")
-      .isLowercase()
-      .withMessage("Name must be a lowercase string")
-      .isFQDN()
-      .withMessage("Name must be a fully qualified domain name")
-      .custom((name) => {
-        const [sub, domain, tld] = name.split(".");
-        if (![`${clickNSDomain}.${parentTLD}`, `${nodleNSDomain}.${parentTLD}`].includes(`${domain}.${tld}`)) {
-          return false;
-        }
-
-        const subHash = keccak256(toUtf8Bytes(sub));
-        if (reservedHashes.includes(subHash)) {
-          return false;
-        }
-
-        return true;
-      })
-      .withMessage("Invalid domain or tld or reserved subdomain")
-      .custom((name) => {
-        const [sub] = name.split(".");
-        if (sub.length < 5) {
-          return false;
-        }
-        return true;
-      })
-      .withMessage(
-        "Current available subdomain names are limited to those with at least 5 characters"
-      ),
-    body("email").isEmail().withMessage("Email must be a valid email address"),
-  ],
-  asyncHandler(async (req: Request, res: Response) => {
-    await checkUserByEmail(req);
-    const result = validationResult(req);
-    if (!result.isEmpty()) {
-      throw new HttpError(
-        result
-          .array()
-          .map((error) => error.msg)
-          .join(", "),
-        400
-      );
-    }
-    const data = matchedData(req);
-
-    const typedData = buildTypedData({
-      name: data.name,
-      email: data.email,
-    });
-
-    res.status(200).send({
-      typedData,
-    });
-  })
-);
-
-app.post(
-  "/name/resolve",
-  body("name")
-    .isFQDN()
-    .withMessage("Name must be a fully qualified domain name")
-    .custom((name) => {
-      const [sub, domain, tld] = name.split(".");
-      if (tld === undefined && ![`${clickNSDomain}.${parentTLD}`, `${nodleNSDomain}.${parentTLD}`].includes(`${domain}.${tld}`)) {
-        return false;
-      }
-      return true;
-    })
-    .withMessage("Invalid domain or tld"),
-  async (req: Request, res: Response) => {
-    try {
-      const result = validationResult(req);
-      if (!result.isEmpty()) {
-        throw new HttpError(
-          result
-            .array()
-            .map((error) => error.msg)
-            .join(", "),
-          400
-        );
-      }
-      const name = matchedData(req).name;
-
-      const parts = name.split(".");
-      const [sub, domain] = parts;
-
-      let owner;
-      if (domain === clickNSDomain) {
-        owner = await clickNameServiceContract.resolve(sub);
-      } else if (domain === nodleNSDomain) {
-        owner = await nodleNameServiceContract.resolve(sub);
-      } else {
-        owner = await l1Provider.resolveName(name);
-      }
-
-      res.status(200).send({
-        owner,
-      });
-    } catch (error) {
-      if (isParsableError(error)) {
-        const decodedError = CLICK_RESOLVER_INTERFACE.parseError(error.data);
-        if (isOffchainLookupError(decodedError)) {
-          const { sender, urls, callData, callbackFunction, extraData } =
-            decodedError.args;
-          res.status(200).send({
-            OffchainLookup: {
-              sender,
-              urls,
-              callData,
-              callbackFunction,
-              extraData,
-            },
-          });
-          return;
-        }
-      }
-      
-      throw error;
-    }
-  }
-);
-
 
 app.use((err: Error, req: Request, res: Response, next: NextFunction): void => {
   if (err instanceof HttpError) {
